+++ src/developer-platform/contracts/developer-security-governance-profile.js
@@ -1,0 +1,53 @@
+export class DeveloperSecurityGovernanceProfile {
+  constructor({
+    programName,
+    objectives = [], principles = [], actors = [], identityTypes = [], aiAgentIdentityFields = [],
+    authorizationFactors = [], authorizationModels = [], accessLifecycleStates = [], aiDelegationFields = [],
+    separationOfDutiesPairs = [], sourceSecurityControls = [], secretProhibitedLocations = [],
+    secretResponseSteps = [], engineeringSecurityDomains = [], aiToolSecuritySteps = [], aiStopConditions = [],
+    threatModelFields = [], policyCharacteristics = [], policyEvaluationFields = [], policyEnforcementPoints = [],
+    riskControlEvidenceFields = [], evidenceLinks = [], reviewApprovalTypes = [], findingFields = [],
+    exceptionLifecycleStates = [], remediationSteps = [], assuranceAreas = [], incidentTriggers = [],
+    supplierGovernanceFields = [], observabilitySignals = [], governanceAreas = [], maturityLevels = [],
+    measures = [], antiPatterns = [], architecturalRules = [], futureCapabilities = [],
+    verifiesEveryIdentity = true, explicitAuthority = true, leastPrivilegeZeroTrustSeparation = true,
+    boundedAiDelegation = true, protectsEngineeringLifecycle = true, preventsSecretUnsafeStorage = true,
+    usesWorkloadIdentitySecretReferences = true, embedsSecurityPrivacyInPavedPaths = true,
+    protectsTenantPropertyIsolation = true, minimizesSensitiveData = true, approvedPolicyAtBoundaries = true,
+    policyAutomationNotAuthority = true, evidenceByDesign = true, independentReviewPreserved = true,
+    exceptionsTemporaryScoped = true, findingsOwnedUntilVerified = true, incidentsEmergencyContinuity = true,
+    governsExternalTools = true, protectsDeveloperPrivacy = true, vendorNeutralTechnologyIndependent = true,
+    replacesArch008OrArch012 = false, infersAuthorityFromCredentials = false, allowsSharedIdentities = false,
+    allowsAiImpersonation = false, allowsAiSelfApproval = false, allowsStandingBroadAccess = false,
+    storesSecretsUnsafely = false, treatsMissingContextAsPermission = false, allowsCrossTenantConvenienceAccess = false,
+    permitsIndefiniteExceptions = false, closesFindingsWithoutEvidence = false,
+    treatsScannerOutputAsFinalRisk = false, usesDeveloperTelemetryForSurveillance = false,
+    reconstructsEvidenceAfterFact = false
+  }) {
+    this.programName = programName;
+    for (const [key, value] of Object.entries({ objectives, principles, actors, identityTypes, aiAgentIdentityFields,
+      authorizationFactors, authorizationModels, accessLifecycleStates, aiDelegationFields, separationOfDutiesPairs,
+      sourceSecurityControls, secretProhibitedLocations, secretResponseSteps, engineeringSecurityDomains,
+      aiToolSecuritySteps, aiStopConditions, threatModelFields, policyCharacteristics, policyEvaluationFields,
+      policyEnforcementPoints, riskControlEvidenceFields, evidenceLinks, reviewApprovalTypes, findingFields,
+      exceptionLifecycleStates, remediationSteps, assuranceAreas, incidentTriggers, supplierGovernanceFields,
+      observabilitySignals, governanceAreas, maturityLevels, measures, antiPatterns, architecturalRules,
+      futureCapabilities })) {
+      this[key] = Object.freeze([...value]);
+    }
+    for (const [key, value] of Object.entries({ verifiesEveryIdentity, explicitAuthority,
+      leastPrivilegeZeroTrustSeparation, boundedAiDelegation, protectsEngineeringLifecycle,
+      preventsSecretUnsafeStorage, usesWorkloadIdentitySecretReferences, embedsSecurityPrivacyInPavedPaths,
+      protectsTenantPropertyIsolation, minimizesSensitiveData, approvedPolicyAtBoundaries,
+      policyAutomationNotAuthority, evidenceByDesign, independentReviewPreserved, exceptionsTemporaryScoped,
+      findingsOwnedUntilVerified, incidentsEmergencyContinuity, governsExternalTools, protectsDeveloperPrivacy,
+      vendorNeutralTechnologyIndependent, replacesArch008OrArch012, infersAuthorityFromCredentials,
+      allowsSharedIdentities, allowsAiImpersonation, allowsAiSelfApproval, allowsStandingBroadAccess,
+      storesSecretsUnsafely, treatsMissingContextAsPermission, allowsCrossTenantConvenienceAccess,
+      permitsIndefiniteExceptions, closesFindingsWithoutEvidence, treatsScannerOutputAsFinalRisk,
+      usesDeveloperTelemetryForSurveillance, reconstructsEvidenceAfterFact })) {
+      this[key] = Boolean(value);
+    }
+    Object.freeze(this);
+  }
+}